requirements.txt
alabaster==0.7.12
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
commonmark==0.9.1
cycler==0.10.0
docutils==0.16
gdevilspie===0.0-svn-
idna==2.10
imagesize==1.2.0
importlib-metadata==3.4.0
iniconfig==1.1.1
Jinja2==2.11.2
kiwisolver==1.3.1
lightning-python==1.2.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.5
packaging==20.8
Pillow==8.1.0
pluggy==0.13.1
py==1.10.0
Pygments==2.7.4
pyln-client==0.9.2
pyparsing==2.4.7
pytest==6.2.1
python-bitcoinlib==0.11.0
python-dateutil==2.8.1
pytz==2020.5
recommonmark==0.7.1
requests==2.25.1
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
zipp==3.4.0
